pyproject.toml
[tool.poetry]
name = "dj-hybrid"
version = "0.1.0"
description = "Playground for hybrid search using Django"
authors = ["Matheus Deister <matheus.deister@vinta.com.br>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
django-ninja = "^1.1.0"
pgvector = "^0.2.5"
openai = "^1.34.0"
transformers = "^4.41.2"
torch = "^2.3.1"
pillow = "^10.3.0"
boto3 = "^1.34.131"
django-storages = "^1.14.3"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "app.app.settings"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.9"
mypy = "^1.10.0"
django-stubs = {extras = ["compatible-mypy"], version = "^5.0.2"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



